// src/core/models/interfaces/management-company.ts
export interface ManagementCompanyDTO {
  id: number
  name: string
  email1: string
  email2: string
  email3: string
  managerName: string
  managerPhone: string
  assitantManagerName: string
  assitantManagerPhone: string
  maintenanceName: string
  maintenancePhone: string
  status: string
  isActive: boolean
  createdAt: string
}